Support opposite fold-outs
Add support for the special case of booklet format with fold-outs in which the two leaves of the same sheet have a fold-out each.

The eight involved pages are rendered as two 4-page sheets glued one to the other with fold-out on each side.

Currently the info about which page is glued to which other page is lost and should be reconstructed manually.